requirements.txt
Pillow==9.1.1
flake8
pytest
